// app/interfaces/VALR.ts
export interface VALRMarketSummery {
  currencyPair: string;
  askPrice: string;
  bidPrice: string;
  lastTradedPrice: string;
  previousClosePrice: string;
  baseVolume: string;
  quoteVolume: string;
  highPrice: string;
  lowPrice: string;
  created: string;
  changeFromPrevious: string;
  markPrice: string;
}